Crate rusoto_directconnect

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud (for example, to Amazon EC2 and Amazon S3) and to Amazon VPC, bypassing Internet service providers in your network path. A connection provides access to all AWS Regions except the China (Beijing) and (China) Ningxia Regions. AWS resources in the China Regions can only be accessed through locations associated with those Regions.

If you're using the service, you're probably looking for DirectConnectClient and DirectConnect.

Structs

AcceptDirectConnectGatewayAssociationProposalRequest

AcceptDirectConnectGatewayAssociationProposalResult

AllocateConnectionOnInterconnectRequest

AllocateHostedConnectionRequest

AllocatePrivateVirtualInterfaceRequest

AllocatePublicVirtualInterfaceRequest

AllocateTransitVirtualInterfaceRequest

AllocateTransitVirtualInterfaceResult

AssociateConnectionWithLagRequest

AssociateHostedConnectionRequest

AssociateVirtualInterfaceRequest

AssociatedGateway
Information about the associated gateway.

BGPPeer
Information about a BGP peer.

ConfirmConnectionRequest

ConfirmConnectionResponse

ConfirmPrivateVirtualInterfaceRequest

ConfirmPrivateVirtualInterfaceResponse

ConfirmPublicVirtualInterfaceRequest

ConfirmPublicVirtualInterfaceResponse

ConfirmTransitVirtualInterfaceRequest

ConfirmTransitVirtualInterfaceResponse

Connection
Information about an AWS Direct Connect connection.

Connections

CreateBGPPeerRequest

CreateBGPPeerResponse

CreateConnectionRequest

CreateDirectConnectGatewayAssociationProposalRequest

CreateDirectConnectGatewayAssociationProposalResult

CreateDirectConnectGatewayAssociationRequest

CreateDirectConnectGatewayAssociationResult

CreateDirectConnectGatewayRequest

CreateDirectConnectGatewayResult

CreateInterconnectRequest

CreateLagRequest

CreatePrivateVirtualInterfaceRequest

CreatePublicVirtualInterfaceRequest

CreateTransitVirtualInterfaceRequest

CreateTransitVirtualInterfaceResult

DeleteBGPPeerRequest

DeleteBGPPeerResponse

DeleteConnectionRequest

DeleteDirectConnectGatewayAssociationProposalRequest

DeleteDirectConnectGatewayAssociationProposalResult

DeleteDirectConnectGatewayAssociationRequest

DeleteDirectConnectGatewayAssociationResult

DeleteDirectConnectGatewayRequest

DeleteDirectConnectGatewayResult

DeleteInterconnectRequest

DeleteInterconnectResponse

DeleteLagRequest

DeleteVirtualInterfaceRequest

DeleteVirtualInterfaceResponse

DescribeConnectionLoaRequest

DescribeConnectionLoaResponse

DescribeConnectionsOnInterconnectRequest

DescribeConnectionsRequest

DescribeDirectConnectGatewayAssociationProposalsRequest

DescribeDirectConnectGatewayAssociationProposalsResult

DescribeDirectConnectGatewayAssociationsRequest

DescribeDirectConnectGatewayAssociationsResult

DescribeDirectConnectGatewayAttachmentsRequest

DescribeDirectConnectGatewayAttachmentsResult

DescribeDirectConnectGatewaysRequest

DescribeDirectConnectGatewaysResult

DescribeHostedConnectionsRequest

DescribeInterconnectLoaRequest

DescribeInterconnectLoaResponse

DescribeInterconnectsRequest

DescribeLagsRequest

DescribeLoaRequest

DescribeTagsRequest

DescribeTagsResponse

DescribeVirtualInterfacesRequest

DirectConnectClient
A client for the AWS Direct Connect API.

DirectConnectGateway
Information about a Direct Connect gateway, which enables you to connect virtual interfaces and virtual private gateway or transit gateways.

DirectConnectGatewayAssociation
Information about an association between a Direct Connect gateway and a virtual private gateway or transit gateway.

DirectConnectGatewayAssociationProposal
Information about the proposal request to attach a virtual private gateway to a Direct Connect gateway.

DirectConnectGatewayAttachment
Information about an attachment between a Direct Connect gateway and a virtual interface.

DisassociateConnectionFromLagRequest

Interconnect
Information about an interconnect.

Interconnects

Lag
Information about a link aggregation group (LAG).

Lags

Loa
Information about a Letter of Authorization - Connecting Facility Assignment (LOA-CFA) for a connection.

Location
Information about an AWS Direct Connect location.

Locations

NewBGPPeer
Information about a new BGP peer.

NewPrivateVirtualInterface
Information about a private virtual interface.

NewPrivateVirtualInterfaceAllocation
Information about a private virtual interface to be provisioned on a connection.

NewPublicVirtualInterface
Information about a public virtual interface.

NewPublicVirtualInterfaceAllocation
Information about a public virtual interface to be provisioned on a connection.

NewTransitVirtualInterface
Information about a transit virtual interface.

NewTransitVirtualInterfaceAllocation
Information about a transit virtual interface to be provisioned on a connection.

ResourceTag
Information about a tag associated with an AWS Direct Connect resource.

RouteFilterPrefix
Information about a route filter prefix that a customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

Tag
Information about a tag.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateDirectConnectGatewayAssociationRequest

UpdateDirectConnectGatewayAssociationResult

UpdateLagRequest

UpdateVirtualInterfaceAttributesRequest

VirtualGateway
Information about a virtual private gateway for a private virtual interface.

VirtualGateways

VirtualInterface
Information about a virtual interface.

VirtualInterfaces

Enums

AcceptDirectConnectGatewayAssociationProposalError
Errors returned by AcceptDirectConnectGatewayAssociationProposal

AllocateConnectionOnInterconnectError
Errors returned by AllocateConnectionOnInterconnect

AllocateHostedConnectionError
Errors returned by AllocateHostedConnection

AllocatePrivateVirtualInterfaceError
Errors returned by AllocatePrivateVirtualInterface

AllocatePublicVirtualInterfaceError
Errors returned by AllocatePublicVirtualInterface

AllocateTransitVirtualInterfaceError
Errors returned by AllocateTransitVirtualInterface

AssociateConnectionWithLagError
Errors returned by AssociateConnectionWithLag

AssociateHostedConnectionError
Errors returned by AssociateHostedConnection

AssociateVirtualInterfaceError
Errors returned by AssociateVirtualInterface

ConfirmConnectionError
Errors returned by ConfirmConnection

ConfirmPrivateVirtualInterfaceError
Errors returned by ConfirmPrivateVirtualInterface

ConfirmPublicVirtualInterfaceError
Errors returned by ConfirmPublicVirtualInterface

ConfirmTransitVirtualInterfaceError
Errors returned by ConfirmTransitVirtualInterface

CreateBGPPeerError
Errors returned by CreateBGPPeer

CreateConnectionError
Errors returned by CreateConnection

CreateDirectConnectGatewayAssociationError
Errors returned by CreateDirectConnectGatewayAssociation

CreateDirectConnectGatewayAssociationProposalError
Errors returned by CreateDirectConnectGatewayAssociationProposal

CreateDirectConnectGatewayError
Errors returned by CreateDirectConnectGateway

CreateInterconnectError
Errors returned by CreateInterconnect

CreateLagError
Errors returned by CreateLag

CreatePrivateVirtualInterfaceError
Errors returned by CreatePrivateVirtualInterface

CreatePublicVirtualInterfaceError
Errors returned by CreatePublicVirtualInterface

CreateTransitVirtualInterfaceError
Errors returned by CreateTransitVirtualInterface

DeleteBGPPeerError
Errors returned by DeleteBGPPeer

DeleteConnectionError
Errors returned by DeleteConnection

DeleteDirectConnectGatewayAssociationError
Errors returned by DeleteDirectConnectGatewayAssociation

DeleteDirectConnectGatewayAssociationProposalError
Errors returned by DeleteDirectConnectGatewayAssociationProposal

DeleteDirectConnectGatewayError
Errors returned by DeleteDirectConnectGateway

DeleteInterconnectError
Errors returned by DeleteInterconnect

DeleteLagError
Errors returned by DeleteLag

DeleteVirtualInterfaceError
Errors returned by DeleteVirtualInterface

DescribeConnectionLoaError
Errors returned by DescribeConnectionLoa

DescribeConnectionsError
Errors returned by DescribeConnections

DescribeConnectionsOnInterconnectError
Errors returned by DescribeConnectionsOnInterconnect

DescribeDirectConnectGatewayAssociationProposalsError
Errors returned by DescribeDirectConnectGatewayAssociationProposals

DescribeDirectConnectGatewayAssociationsError
Errors returned by DescribeDirectConnectGatewayAssociations

DescribeDirectConnectGatewayAttachmentsError
Errors returned by DescribeDirectConnectGatewayAttachments

DescribeDirectConnectGatewaysError
Errors returned by DescribeDirectConnectGateways

DescribeHostedConnectionsError
Errors returned by DescribeHostedConnections

DescribeInterconnectLoaError
Errors returned by DescribeInterconnectLoa

DescribeInterconnectsError
Errors returned by DescribeInterconnects

DescribeLagsError
Errors returned by DescribeLags

DescribeLoaError
Errors returned by DescribeLoa

DescribeLocationsError
Errors returned by DescribeLocations

DescribeTagsError
Errors returned by DescribeTags

DescribeVirtualGatewaysError
Errors returned by DescribeVirtualGateways

DescribeVirtualInterfacesError
Errors returned by DescribeVirtualInterfaces

DisassociateConnectionFromLagError
Errors returned by DisassociateConnectionFromLag

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateDirectConnectGatewayAssociationError
Errors returned by UpdateDirectConnectGatewayAssociation

UpdateLagError
Errors returned by UpdateLag

UpdateVirtualInterfaceAttributesError
Errors returned by UpdateVirtualInterfaceAttributes

Traits

DirectConnect
Trait representing the capabilities of the AWS Direct Connect API. AWS Direct Connect clients implement this trait.