Struct rusoto_directconnect::Lag

pub struct Lag {
    pub allows_hosted_connections: Option<bool>,
    pub aws_device: Option<String>,
    pub connections: Option<Vec<Connection>>,
    pub connections_bandwidth: Option<String>,
    pub lag_id: Option<String>,
    pub lag_name: Option<String>,
    pub lag_state: Option<String>,
    pub location: Option<String>,
    pub minimum_links: Option<i64>,
    pub number_of_connections: Option<i64>,
    pub owner_account: Option<String>,
    pub region: Option<String>,
}

Describes a link aggregation group (LAG). A LAG is a connection that uses the Link Aggregation Control Protocol (LACP) to logically aggregate a bundle of physical connections. Like an interconnect, it can host other connections. All connections in a LAG must terminate on the same physical AWS Direct Connect endpoint, and must be the same bandwidth.

Fields
allows_hosted_connections: Option<bool>

Indicates whether the LAG can host other connections.

This is intended for use by AWS Direct Connect partners only.

aws_device: Option<String>

The AWS Direct Connection endpoint that hosts the LAG.

connections: Option<Vec<Connection>>

A list of connections bundled by this LAG.

connections_bandwidth: Option<String>

The individual bandwidth of the physical connections bundled by the LAG.

Available values: 1Gbps, 10Gbps

lag_id: Option<String> lag_name: Option<String>

The name of the LAG.

lag_state: Option<String> location: Option<String> minimum_links: Option<i64>

The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down. This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

number_of_connections: Option<i64>

The number of physical connections bundled by the LAG, up to a maximum of 10.

owner_account: Option<String>

The owner of the LAG.

region: Option<String>