Struct rusoto_directconnect::VirtualInterface

pub struct VirtualInterface {
    pub address_family: Option<String>,
    pub amazon_address: Option<String>,
    pub amazon_side_asn: Option<i64>,
    pub asn: Option<i64>,
    pub auth_key: Option<String>,
    pub aws_device_v2: Option<String>,
    pub bgp_peers: Option<Vec<BGPPeer>>,
    pub connection_id: Option<String>,
    pub customer_address: Option<String>,
    pub customer_router_config: Option<String>,
    pub direct_connect_gateway_id: Option<String>,
    pub jumbo_frame_capable: Option<bool>,
    pub location: Option<String>,
    pub mtu: Option<i64>,
    pub owner_account: Option<String>,
    pub region: Option<String>,
    pub route_filter_prefixes: Option<Vec<RouteFilterPrefix>>,
    pub virtual_gateway_id: Option<String>,
    pub virtual_interface_id: Option<String>,
    pub virtual_interface_name: Option<String>,
    pub virtual_interface_state: Option<String>,
    pub virtual_interface_type: Option<String>,
    pub vlan: Option<i64>,
}

Information about a virtual interface.

Fields
address_family: Option<String>

The address family for the BGP peer.

amazon_address: Option<String>

The IP address assigned to the Amazon interface.

amazon_side_asn: Option<i64>

The autonomous system number (ASN) for the Amazon side of the connection.

asn: Option<i64>

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

auth_key: Option<String>

The authentication key for BGP configuration.

aws_device_v2: Option<String>

The Direct Connect endpoint on which the virtual interface terminates.

bgp_peers: Option<Vec<BGPPeer>>

The BGP peers configured on this virtual interface.

connection_id: Option<String>

The ID of the connection.

customer_address: Option<String>

The IP address assigned to the customer interface.

customer_router_config: Option<String>

The customer router configuration.

direct_connect_gateway_id: Option<String>

The ID of the Direct Connect gateway.

jumbo_frame_capable: Option<bool>

Indicates whether jumbo frames (9001 MTU) are supported.

location: Option<String>

The location of the connection.

mtu: Option<i64>

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

owner_account: Option<String>

The ID of the AWS account that owns the virtual interface.

region: Option<String>

The AWS Region where the virtual interface is located.

route_filter_prefixes: Option<Vec<RouteFilterPrefix>>

The routes to be advertised to the AWS network in this Region. Applies to public virtual interfaces.

virtual_gateway_id: Option<String>

The ID of the virtual private gateway. Applies only to private virtual interfaces.

virtual_interface_id: Option<String>

The ID of the virtual interface.

virtual_interface_name: Option<String>

The name of the virtual interface assigned by the customer network.

virtual_interface_state: Option<String>

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

virtual_interface_type: Option<String>

The type of virtual interface. The possible values are private and public.

vlan: Option<i64>

The ID of the VLAN.