Type Definition rusoto_directconnect::BGPPeerState

type BGPPeerState = String;

The state of the BGP peer.

Verifying: The BGP peering addresses or ASN require validation before the BGP peer can be created. This state only applies to BGP peers on a public virtual interface.

Pending: The BGP peer has been created, and is in this state until it is ready to be established.

Available: The BGP peer can be established.

Deleting: The BGP peer is in the process of being deleted.

Deleted: The BGP peer has been deleted and cannot be established.