Type Definition rusoto_directconnect::VirtualInterfaceState

type VirtualInterfaceState = String;

State of the virtual interface.

Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

Available: A virtual interface that is able to forward traffic.

Down: A virtual interface that is BGP down.

Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

Deleted: A virtual interface that cannot forward traffic.

Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the 'Confirming' state is deleted by the virtual interface owner, the virtual interface will enter the 'Rejected' state.