Crate rusoto_directconnect

AWS Direct Connect

If you're using the service, you're probably looking for DirectConnectClient and DirectConnect.

Structs

AllocateConnectionOnInterconnectRequest
Container for the parameters to the AllocateConnectionOnInterconnect operation.

AllocateHostedConnectionRequest
Container for the parameters to theHostedConnection operation.

AllocatePrivateVirtualInterfaceRequest
Container for the parameters to the AllocatePrivateVirtualInterface operation.

AllocatePublicVirtualInterfaceRequest
Container for the parameters to the AllocatePublicVirtualInterface operation.

AssociateConnectionWithLagRequest
Container for the parameters to the AssociateConnectionWithLag operation.

AssociateHostedConnectionRequest
Container for the parameters to the AssociateHostedConnection operation.

AssociateVirtualInterfaceRequest
Container for the parameters to the AssociateVirtualInterface operation.

BGPPeer
A structure containing information about a BGP peer.

ConfirmConnectionRequest
Container for the parameters to the ConfirmConnection operation.

ConfirmConnectionResponse
The response received when ConfirmConnection is called.

ConfirmPrivateVirtualInterfaceRequest
Container for the parameters to the ConfirmPrivateVirtualInterface operation.

ConfirmPrivateVirtualInterfaceResponse
The response received when ConfirmPrivateVirtualInterface is called.

ConfirmPublicVirtualInterfaceRequest
Container for the parameters to the ConfirmPublicVirtualInterface operation.

ConfirmPublicVirtualInterfaceResponse
The response received when ConfirmPublicVirtualInterface is called.

Connection
A connection represents the physical network connection between the AWS Direct Connect location and the customer.

Connections
A structure containing a list of connections.

CreateBGPPeerRequest
Container for the parameters to the CreateBGPPeer operation.

CreateBGPPeerResponse
The response received when CreateBGPPeer is called.

CreateConnectionRequest
Container for the parameters to the CreateConnection operation.

CreateInterconnectRequest
Container for the parameters to the CreateInterconnect operation.

CreateLagRequest
Container for the parameters to the CreateLag operation.

CreatePrivateVirtualInterfaceRequest
Container for the parameters to the CreatePrivateVirtualInterface operation.

CreatePublicVirtualInterfaceRequest
Container for the parameters to the CreatePublicVirtualInterface operation.

DeleteBGPPeerRequest
Container for the parameters to the DeleteBGPPeer operation.

DeleteBGPPeerResponse
The response received when DeleteBGPPeer is called.

DeleteConnectionRequest
Container for the parameters to the DeleteConnection operation.

DeleteInterconnectRequest
Container for the parameters to the DeleteInterconnect operation.

DeleteInterconnectResponse
The response received when DeleteInterconnect is called.

DeleteLagRequest
Container for the parameters to the DeleteLag operation.

DeleteVirtualInterfaceRequest
Container for the parameters to the DeleteVirtualInterface operation.

DeleteVirtualInterfaceResponse
The response received when DeleteVirtualInterface is called.

DescribeConnectionLoaRequest
Container for the parameters to the DescribeConnectionLoa operation.

DescribeConnectionLoaResponse
The response received when DescribeConnectionLoa is called.

DescribeConnectionsOnInterconnectRequest
Container for the parameters to the DescribeConnectionsOnInterconnect operation.

DescribeConnectionsRequest
Container for the parameters to the DescribeConnections operation.

DescribeHostedConnectionsRequest
Container for the parameters to the DescribeHostedConnections operation.

DescribeInterconnectLoaRequest
Container for the parameters to the DescribeInterconnectLoa operation.

DescribeInterconnectLoaResponse
The response received when DescribeInterconnectLoa is called.

DescribeInterconnectsRequest
Container for the parameters to the DescribeInterconnects operation.

DescribeLagsRequest
Container for the parameters to the DescribeLags operation.

DescribeLoaRequest
Container for the parameters to the DescribeLoa operation.

DescribeTagsRequest
Container for the parameters to the DescribeTags operation.

DescribeTagsResponse
The response received when DescribeTags is called.

DescribeVirtualInterfacesRequest
Container for the parameters to the DescribeVirtualInterfaces operation.

DirectConnectClient
A client for the AWS Direct Connect API.

DisassociateConnectionFromLagRequest
Container for the parameters to the DisassociateConnectionFromLag operation.

Interconnect
An interconnect is a connection that can host other connections.

Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services.

The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

Interconnects
A structure containing a list of interconnects.

Lag
Describes a link aggregation group (LAG). A LAG is a connection that uses the Link Aggregation Control Protocol (LACP) to logically aggregate a bundle of physical connections. Like an interconnect, it can host other connections. All connections in a LAG must terminate on the same physical AWS Direct Connect endpoint, and must be the same bandwidth.

Lags
A structure containing a list of LAGs.

Loa
A structure containing the Letter of Authorization - Connecting Facility Assignment (LOA-CFA) for a connection.

Location
An AWS Direct Connect location where connections and interconnects can be requested.

Locations
A location is a network facility where AWS Direct Connect routers are available to be connected. Generally, these are colocation hubs where many network providers have equipment, and where cross connects can be delivered. Locations include a name and facility code, and must be provided when creating a connection.

NewBGPPeer
A structure containing information about a new BGP peer.

NewPrivateVirtualInterface
A structure containing information about a new private virtual interface.

NewPrivateVirtualInterfaceAllocation
A structure containing information about a private virtual interface that will be provisioned on a connection.

NewPublicVirtualInterface
A structure containing information about a new public virtual interface.

NewPublicVirtualInterfaceAllocation
A structure containing information about a public virtual interface that will be provisioned on a connection.

ResourceTag
The tags associated with a Direct Connect resource.

RouteFilterPrefix
A route filter prefix that the customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

Tag
Information about a tag.

TagResourceRequest
Container for the parameters to the TagResource operation.

TagResourceResponse
The response received when TagResource is called.

UntagResourceRequest
Container for the parameters to the UntagResource operation.

UntagResourceResponse
The response received when UntagResource is called.

UpdateLagRequest
Container for the parameters to the UpdateLag operation.

VirtualGateway
You can create one or more AWS Direct Connect private virtual interfaces linking to your virtual private gateway.

Virtual private gateways can be managed using the Amazon Virtual Private Cloud (Amazon VPC) console or the Amazon EC2 CreateVpnGateway action.

VirtualGateways
A structure containing a list of virtual private gateways.

VirtualInterface
A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

VirtualInterfaces
A structure containing a list of virtual interfaces.

Enums

AllocateConnectionOnInterconnectError
Errors returned by AllocateConnectionOnInterconnect

AllocateHostedConnectionError
Errors returned by AllocateHostedConnection

AllocatePrivateVirtualInterfaceError
Errors returned by AllocatePrivateVirtualInterface

AllocatePublicVirtualInterfaceError
Errors returned by AllocatePublicVirtualInterface

AssociateConnectionWithLagError
Errors returned by AssociateConnectionWithLag

AssociateHostedConnectionError
Errors returned by AssociateHostedConnection

AssociateVirtualInterfaceError
Errors returned by AssociateVirtualInterface

ConfirmConnectionError
Errors returned by ConfirmConnection

ConfirmPrivateVirtualInterfaceError
Errors returned by ConfirmPrivateVirtualInterface

ConfirmPublicVirtualInterfaceError
Errors returned by ConfirmPublicVirtualInterface

CreateBGPPeerError
Errors returned by CreateBGPPeer

CreateConnectionError
Errors returned by CreateConnection

CreateInterconnectError
Errors returned by CreateInterconnect

CreateLagError
Errors returned by CreateLag

CreatePrivateVirtualInterfaceError
Errors returned by CreatePrivateVirtualInterface

CreatePublicVirtualInterfaceError
Errors returned by CreatePublicVirtualInterface

DeleteBGPPeerError
Errors returned by DeleteBGPPeer

DeleteConnectionError
Errors returned by DeleteConnection

DeleteInterconnectError
Errors returned by DeleteInterconnect

DeleteLagError
Errors returned by DeleteLag

DeleteVirtualInterfaceError
Errors returned by DeleteVirtualInterface

DescribeConnectionLoaError
Errors returned by DescribeConnectionLoa

DescribeConnectionsError
Errors returned by DescribeConnections

DescribeConnectionsOnInterconnectError
Errors returned by DescribeConnectionsOnInterconnect

DescribeHostedConnectionsError
Errors returned by DescribeHostedConnections

DescribeInterconnectLoaError
Errors returned by DescribeInterconnectLoa

DescribeInterconnectsError
Errors returned by DescribeInterconnects

DescribeLagsError
Errors returned by DescribeLags

DescribeLoaError
Errors returned by DescribeLoa

DescribeLocationsError
Errors returned by DescribeLocations

DescribeTagsError
Errors returned by DescribeTags

DescribeVirtualGatewaysError
Errors returned by DescribeVirtualGateways

DescribeVirtualInterfacesError
Errors returned by DescribeVirtualInterfaces

DisassociateConnectionFromLagError
Errors returned by DisassociateConnectionFromLag

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateLagError
Errors returned by UpdateLag

Traits

DirectConnect
Trait representing the capabilities of the AWS Direct Connect API. AWS Direct Connect clients implement this trait.

Type Definitions

ASN
The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

Example: 65000

AddressFamily

Indicates the address family for the BGP peer.

ipv4: IPv4 address family

ipv6: IPv6 address family

AmazonAddress
IP address assigned to the Amazon interface.

Example: 192.168.1.1/30 or 2001:db8::1/125

AwsDevice
An abstract ID for the physical Direct Connect endpoint.

Example: EQC50-abcdef123456

BGPAuthKey
The authentication key for BGP configuration.

Example: asdf34example

BGPPeerList
A list of the BGP peers configured on this virtual interface.

BGPPeerState

The state of the BGP peer.

Verifying: The BGP peering addresses or ASN require validation before the BGP peer can be created. This state only applies to BGP peers on a public virtual interface.

Pending: The BGP peer has been created, and is in this state until it is ready to be established.

Available: The BGP peer can be established.

Deleting: The BGP peer is in the process of being deleted.

Deleted: The BGP peer has been deleted and cannot be established.

BGPStatus

The Up/Down state of the BGP peer.

Up: The BGP peer is established.

Down: The BGP peer is down.

Bandwidth
Bandwidth of the connection.

Example: 1Gbps

Default: None

BooleanFlag

CIDR

ConnectionId
The ID of the connection. This field is also used as the ID type for operations that use multiple connection types (LAG, interconnect, and/or connection).

Example: dxcon-fg5678gh

Default: None

ConnectionList
A list of connections.

ConnectionName
The name of the connection.

Example: "My Connection to AWS"

Default: None

ConnectionState

State of the connection.

Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

Pending: The connection has been approved, and is being initialized.

Available: The network link is up, and the connection is ready for use.

Down: The network link is down.

Deleting: The connection is in the process of being deleted.

Deleted: The connection has been deleted.

Rejected: A hosted connection in the 'Ordering' state will enter the 'Rejected' state if it is deleted by the end customer.

Count

CustomerAddress
IP address assigned to the customer interface.

Example: 192.168.1.2/30 or 2001:db8::2/125

ErrorMessage

InterconnectId
The ID of the interconnect.

Example: dxcon-abc123

InterconnectList
A list of interconnects.

InterconnectName
The name of the interconnect.

Example: "1G Interconnect to AWS"

InterconnectState

State of the interconnect.

Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

Pending: The interconnect has been approved, and is being initialized.

Available: The network link is up, and the interconnect is ready for use.

Down: The network link is down.

Deleting: The interconnect is in the process of being deleted.

Deleted: The interconnect has been deleted.

LagId
The ID of the LAG.

Example: dxlag-fg5678gh

LagList
A list of LAGs.

LagName

LagState

The state of the LAG.

Requested: The initial state of a LAG. The LAG stays in the requested state until the Letter of Authorization (LOA) is available.

Pending: The LAG has been approved, and is being initialized.

Available: The network link is established, and the LAG is ready for use.

Down: The network link is down.

Deleting: The LAG is in the process of being deleted.

Deleted: The LAG has been deleted.

LoaContent
The binary contents of the LOA-CFA document.

LoaContentType
A standard media type indicating the content type of the LOA-CFA document. Currently, the only supported value is "application/pdf".

Default: application/pdf

LoaIssueTime

LocationCode
Where the connection is located.

Example: EqSV5

Default: None

LocationList

LocationName

OwnerAccount

PartnerName

ProviderName

Region
The AWS region where the connection is located.

Example: us-east-1

Default: None

ResourceArn

ResourceArnList

ResourceTagList

RouteFilterPrefixList
A list of routes to be advertised to the AWS network in this region (public virtual interface).

RouterConfig

TagKey

TagKeyList

TagList

TagValue

VLAN
The VLAN ID.

Example: 101

VirtualGatewayId
The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces.

Example: vgw-123er56

VirtualGatewayList
A list of virtual private gateways.

VirtualGatewayState

State of the virtual private gateway.

Pending: This is the initial state after calling CreateVpnGateway.

Available: Ready for use by a private virtual interface.

Deleting: This is the initial state after calling DeleteVpnGateway.

Deleted: In this state, a private virtual interface is unable to send traffic over this gateway.

VirtualInterfaceId
The ID of the virtual interface.

Example: dxvif-123dfg56

Default: None

VirtualInterfaceList
A list of virtual interfaces.

VirtualInterfaceName
The name of the virtual interface assigned by the customer.

Example: "My VPC"

VirtualInterfaceState

State of the virtual interface.

Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

Available: A virtual interface that is able to forward traffic.

Down: A virtual interface that is BGP down.

Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

Deleted: A virtual interface that cannot forward traffic.

Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the 'Confirming' state is deleted by the virtual interface owner, the virtual interface will enter the 'Rejected' state.

VirtualInterfaceType
The type of virtual interface.

Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.)