Crate rusoto_directconnect

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud (for example, to Amazon Elastic Compute Cloud (Amazon EC2) and Amazon Simple Storage Service (Amazon S3)) and to Amazon Virtual Private Cloud (Amazon VPC), bypassing Internet service providers in your network path. An AWS Direct Connect location provides access to AWS in the region it is associated with, as well as access to other US regions. For example, you can provision a single connection to any AWS Direct Connect location in the US and use it to access public AWS services in all US Regions and AWS GovCloud (US).

If you're using the service, you're probably looking for DirectConnectClient and DirectConnect.

Structs

AllocateConnectionOnInterconnectRequest
Container for the parameters to the AllocateConnectionOnInterconnect operation.

AllocateHostedConnectionRequest
Container for the parameters to theHostedConnection operation.

AllocatePrivateVirtualInterfaceRequest
Container for the parameters to the AllocatePrivateVirtualInterface operation.

AllocatePublicVirtualInterfaceRequest
Container for the parameters to the AllocatePublicVirtualInterface operation.

AssociateConnectionWithLagRequest
Container for the parameters to the AssociateConnectionWithLag operation.

AssociateHostedConnectionRequest
Container for the parameters to the AssociateHostedConnection operation.

AssociateVirtualInterfaceRequest
Container for the parameters to the AssociateVirtualInterface operation.

BGPPeer
A structure containing information about a BGP peer.

ConfirmConnectionRequest
Container for the parameters to the ConfirmConnection operation.

ConfirmConnectionResponse
The response received when ConfirmConnection is called.

ConfirmPrivateVirtualInterfaceRequest
Container for the parameters to the ConfirmPrivateVirtualInterface operation.

ConfirmPrivateVirtualInterfaceResponse
The response received when ConfirmPrivateVirtualInterface is called.

ConfirmPublicVirtualInterfaceRequest
Container for the parameters to the ConfirmPublicVirtualInterface operation.

ConfirmPublicVirtualInterfaceResponse
The response received when ConfirmPublicVirtualInterface is called.

Connection
A connection represents the physical network connection between the AWS Direct Connect location and the customer.

Connections
A structure containing a list of connections.

CreateBGPPeerRequest
Container for the parameters to the CreateBGPPeer operation.

CreateBGPPeerResponse
The response received when CreateBGPPeer is called.

CreateConnectionRequest
Container for the parameters to the CreateConnection operation.

CreateDirectConnectGatewayAssociationRequest
Container for the parameters to the CreateDirectConnectGatewayAssociation operation.

CreateDirectConnectGatewayAssociationResult
Container for the response from the CreateDirectConnectGatewayAssociation API call

CreateDirectConnectGatewayRequest
Container for the parameters to the CreateDirectConnectGateway operation.

CreateDirectConnectGatewayResult
Container for the response from the CreateDirectConnectGateway API call

CreateInterconnectRequest
Container for the parameters to the CreateInterconnect operation.

CreateLagRequest
Container for the parameters to the CreateLag operation.

CreatePrivateVirtualInterfaceRequest
Container for the parameters to the CreatePrivateVirtualInterface operation.

CreatePublicVirtualInterfaceRequest
Container for the parameters to the CreatePublicVirtualInterface operation.

DeleteBGPPeerRequest
Container for the parameters to the DeleteBGPPeer operation.

DeleteBGPPeerResponse
The response received when DeleteBGPPeer is called.

DeleteConnectionRequest
Container for the parameters to the DeleteConnection operation.

DeleteDirectConnectGatewayAssociationRequest
Container for the parameters to the DeleteDirectConnectGatewayAssociation operation.

DeleteDirectConnectGatewayAssociationResult
Container for the response from the DeleteDirectConnectGatewayAssociation API call

DeleteDirectConnectGatewayRequest
Container for the parameters to the DeleteDirectConnectGateway operation.

DeleteDirectConnectGatewayResult
Container for the response from the DeleteDirectConnectGateway API call

DeleteInterconnectRequest
Container for the parameters to the DeleteInterconnect operation.

DeleteInterconnectResponse
The response received when DeleteInterconnect is called.

DeleteLagRequest
Container for the parameters to the DeleteLag operation.

DeleteVirtualInterfaceRequest
Container for the parameters to the DeleteVirtualInterface operation.

DeleteVirtualInterfaceResponse
The response received when DeleteVirtualInterface is called.

DescribeConnectionLoaRequest
Container for the parameters to the DescribeConnectionLoa operation.

DescribeConnectionLoaResponse
The response received when DescribeConnectionLoa is called.

DescribeConnectionsOnInterconnectRequest
Container for the parameters to the DescribeConnectionsOnInterconnect operation.

DescribeConnectionsRequest
Container for the parameters to the DescribeConnections operation.

DescribeDirectConnectGatewayAssociationsRequest
Container for the parameters to the DescribeDirectConnectGatewayAssociations operation.

DescribeDirectConnectGatewayAssociationsResult
Container for the response from the DescribeDirectConnectGatewayAssociations API call

DescribeDirectConnectGatewayAttachmentsRequest
Container for the parameters to the DescribeDirectConnectGatewayAttachments operation.

DescribeDirectConnectGatewayAttachmentsResult
Container for the response from the DescribeDirectConnectGatewayAttachments API call

DescribeDirectConnectGatewaysRequest
Container for the parameters to the DescribeDirectConnectGateways operation.

DescribeDirectConnectGatewaysResult
Container for the response from the DescribeDirectConnectGateways API call

DescribeHostedConnectionsRequest
Container for the parameters to the DescribeHostedConnections operation.

DescribeInterconnectLoaRequest
Container for the parameters to the DescribeInterconnectLoa operation.

DescribeInterconnectLoaResponse
The response received when DescribeInterconnectLoa is called.

DescribeInterconnectsRequest
Container for the parameters to the DescribeInterconnects operation.

DescribeLagsRequest
Container for the parameters to the DescribeLags operation.

DescribeLoaRequest
Container for the parameters to the DescribeLoa operation.

DescribeTagsRequest
Container for the parameters to the DescribeTags operation.

DescribeTagsResponse
The response received when DescribeTags is called.

DescribeVirtualInterfacesRequest
Container for the parameters to the DescribeVirtualInterfaces operation.

DirectConnectClient
A client for the AWS Direct Connect API.

DirectConnectGateway
A direct connect gateway is an intermediate object that enables you to connect virtual interfaces and virtual private gateways.

DirectConnectGatewayAssociation
The association between a direct connect gateway and virtual private gateway.

DirectConnectGatewayAttachment
The association between a direct connect gateway and virtual interface.

DisassociateConnectionFromLagRequest
Container for the parameters to the DisassociateConnectionFromLag operation.

Interconnect
An interconnect is a connection that can host other connections.

Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services.

The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

Interconnects
A structure containing a list of interconnects.

Lag
Describes a link aggregation group (LAG). A LAG is a connection that uses the Link Aggregation Control Protocol (LACP) to logically aggregate a bundle of physical connections. Like an interconnect, it can host other connections. All connections in a LAG must terminate on the same physical AWS Direct Connect endpoint, and must be the same bandwidth.

Lags
A structure containing a list of LAGs.

Loa
A structure containing the Letter of Authorization - Connecting Facility Assignment (LOA-CFA) for a connection.

Location
An AWS Direct Connect location where connections and interconnects can be requested.

Locations
A location is a network facility where AWS Direct Connect routers are available to be connected. Generally, these are colocation hubs where many network providers have equipment, and where cross connects can be delivered. Locations include a name and facility code, and must be provided when creating a connection.

NewBGPPeer
A structure containing information about a new BGP peer.

NewPrivateVirtualInterface
A structure containing information about a new private virtual interface.

NewPrivateVirtualInterfaceAllocation
A structure containing information about a private virtual interface that will be provisioned on a connection.

NewPublicVirtualInterface
A structure containing information about a new public virtual interface.

NewPublicVirtualInterfaceAllocation
A structure containing information about a public virtual interface that will be provisioned on a connection.

ResourceTag
The tags associated with a Direct Connect resource.

RouteFilterPrefix
A route filter prefix that the customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

Tag
Information about a tag.

TagResourceRequest
Container for the parameters to the TagResource operation.

TagResourceResponse
The response received when TagResource is called.

UntagResourceRequest
Container for the parameters to the UntagResource operation.

UntagResourceResponse
The response received when UntagResource is called.

UpdateLagRequest
Container for the parameters to the UpdateLag operation.

VirtualGateway
You can create one or more AWS Direct Connect private virtual interfaces linking to your virtual private gateway.

Virtual private gateways can be managed using the Amazon Virtual Private Cloud (Amazon VPC) console or the Amazon EC2 CreateVpnGateway action.

VirtualGateways
A structure containing a list of virtual private gateways.

VirtualInterface
A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

VirtualInterfaces
A structure containing a list of virtual interfaces.

Enums

AllocateConnectionOnInterconnectError
Errors returned by AllocateConnectionOnInterconnect

AllocateHostedConnectionError
Errors returned by AllocateHostedConnection

AllocatePrivateVirtualInterfaceError
Errors returned by AllocatePrivateVirtualInterface

AllocatePublicVirtualInterfaceError
Errors returned by AllocatePublicVirtualInterface

AssociateConnectionWithLagError
Errors returned by AssociateConnectionWithLag

AssociateHostedConnectionError
Errors returned by AssociateHostedConnection

AssociateVirtualInterfaceError
Errors returned by AssociateVirtualInterface

ConfirmConnectionError
Errors returned by ConfirmConnection

ConfirmPrivateVirtualInterfaceError
Errors returned by ConfirmPrivateVirtualInterface

ConfirmPublicVirtualInterfaceError
Errors returned by ConfirmPublicVirtualInterface

CreateBGPPeerError
Errors returned by CreateBGPPeer

CreateConnectionError
Errors returned by CreateConnection

CreateDirectConnectGatewayAssociationError
Errors returned by CreateDirectConnectGatewayAssociation

CreateDirectConnectGatewayError
Errors returned by CreateDirectConnectGateway

CreateInterconnectError
Errors returned by CreateInterconnect

CreateLagError
Errors returned by CreateLag

CreatePrivateVirtualInterfaceError
Errors returned by CreatePrivateVirtualInterface

CreatePublicVirtualInterfaceError
Errors returned by CreatePublicVirtualInterface

DeleteBGPPeerError
Errors returned by DeleteBGPPeer

DeleteConnectionError
Errors returned by DeleteConnection

DeleteDirectConnectGatewayAssociationError
Errors returned by DeleteDirectConnectGatewayAssociation

DeleteDirectConnectGatewayError
Errors returned by DeleteDirectConnectGateway

DeleteInterconnectError
Errors returned by DeleteInterconnect

DeleteLagError
Errors returned by DeleteLag

DeleteVirtualInterfaceError
Errors returned by DeleteVirtualInterface

DescribeConnectionLoaError
Errors returned by DescribeConnectionLoa

DescribeConnectionsError
Errors returned by DescribeConnections

DescribeConnectionsOnInterconnectError
Errors returned by DescribeConnectionsOnInterconnect

DescribeDirectConnectGatewayAssociationsError
Errors returned by DescribeDirectConnectGatewayAssociations

DescribeDirectConnectGatewayAttachmentsError
Errors returned by DescribeDirectConnectGatewayAttachments

DescribeDirectConnectGatewaysError
Errors returned by DescribeDirectConnectGateways

DescribeHostedConnectionsError
Errors returned by DescribeHostedConnections

DescribeInterconnectLoaError
Errors returned by DescribeInterconnectLoa

DescribeInterconnectsError
Errors returned by DescribeInterconnects

DescribeLagsError
Errors returned by DescribeLags

DescribeLoaError
Errors returned by DescribeLoa

DescribeLocationsError
Errors returned by DescribeLocations

DescribeTagsError
Errors returned by DescribeTags

DescribeVirtualGatewaysError
Errors returned by DescribeVirtualGateways

DescribeVirtualInterfacesError
Errors returned by DescribeVirtualInterfaces

DisassociateConnectionFromLagError
Errors returned by DisassociateConnectionFromLag

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateLagError
Errors returned by UpdateLag

Traits

DirectConnect
Trait representing the capabilities of the AWS Direct Connect API. AWS Direct Connect clients implement this trait.