Struct rusoto_directconnect::DirectConnectClient

pub struct DirectConnectClient { /* fields omitted */ }

A client for the AWS Direct Connect API.

Methods

impl DirectConnectClient

pub fn new(region: Region) -> DirectConnectClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DirectConnectClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,